# small optimisations in Log2/Sqrt, rework of multiplication of fp128 t…