small optimisations in Log2/Sqrt
rework of multiplication of fp128 to not use the Int<256>

before:
```
BM_Abs<fetch::fixed_point::fp128_t>/1000000            717 ns        712 ns     873582
BM_exp<fetch::fixed_point::fp128_t>/1000000        3540000 ns    3518738 ns        193
BM_Log<fetch::fixed_point::fp128_t>/1000000        6350214 ns    6352381 ns        109
BM_Log2<fetch::fixed_point::fp128_t>/1000000       5154450 ns    5191308 ns        100
BM_Log10<fetch::fixed_point::fp128_t>/1000000      6497776 ns    6534259 ns         99
BM_Pow<1, fetch::fixed_point::fp128_t>/1000000    10496420 ns   10533411 ns         65
BM_Pow<2, fetch::fixed_point::fp128_t>/1000000    11373177 ns   11442274 ns         57
BM_Pow<3, fetch::fixed_point::fp128_t>/1000000    11470462 ns   11544433 ns         64
BM_Pow<10, fetch::fixed_point::fp128_t>/1000000   11238363 ns   11305321 ns         63
```

after
```
BM_Abs<fetch::fixed_point::fp128_t>/1000000            692 ns        695 ns     885724
BM_exp<fetch::fixed_point::fp128_t>/1000000        3123120 ns    3123080 ns        217
BM_Log<fetch::fixed_point::fp128_t>/1000000        5995180 ns    6033377 ns        118
BM_Log2<fetch::fixed_point::fp128_t>/1000000       4294665 ns    4319492 ns        165
BM_Log10<fetch::fixed_point::fp128_t>/1000000      6050873 ns    6091777 ns        116
BM_Pow<1, fetch::fixed_point::fp128_t>/1000000     9945404 ns   10003643 ns         73
BM_Pow<2, fetch::fixed_point::fp128_t>/1000000     9992760 ns   10050175 ns         70
BM_Pow<3, fetch::fixed_point::fp128_t>/1000000     9729734 ns    9780734 ns         72
BM_Pow<10, fetch::fixed_point::fp128_t>/1000000    9289753 ns    9328825 ns         74
```